Fix score read width in PlayerProperties handler (LE uint32, not byte)

+++ Sources/Client/NetClient.cpp
@@ -1472,7 +1472,7 @@ namespace spades {
 					int grenades = r.ReadByte();
 					int clip = r.ReadByte();
 					int reserve = r.ReadByte();
-					int score = r.ReadByte();
+					int score = r.ReadInt();
 
 					Player& p = GetPlayer(pId);
 					Weapon& w = p.GetWeapon();
